Re-write handling of command pass through
To accomodate environments where we wish to issue a command with
a Virtual Machine name on the end.. e.g.
up <vm name> or ssh <vm name> in a multiple VM vagrantfile, we
have to be able to pass through an extra block of stuff.

Now calling the command is no longer a switch with an if-else chain, we
simply call runCommand and pass it an argument chain.